+++ arch/x86/include/asm/pkvm.h
@@ -24,6 +24,7 @@
 #define PKVM_HC_TLB_REMOTE_FLUSH_RANGE	14
 #define PKVM_HC_SET_MMIO_VE		15
 #define PKVM_HC_ADD_PTDEV		16
+#define PKVM_HC_SUBMIT_QI       17
 
 #define PKVM_HC_DUMP_DMAR_TR_STRUCT	20
 #define PKVM_HC_DUMP_DOMAIN_PGT		21
