+++ arch/x86/kvm/vmx/pkvm/hyp/iommu.c
@@ -8,6 +8,7 @@
 #include "pkvm_hyp.h"
 #include "gfp.h"
 #include "memory.h"
+#include "mem_protect.h"
 #include "mmu.h"
 #include "ept.h"
 #include "pgtable.h"
@@ -281,6 +282,9 @@ static int quiesce_qi(struct pkvm_iommu *iommu)
 static int create_qi_desc(struct pkvm_iommu *iommu)
 {
 	struct q_inval *qi = &iommu->qi;
+#ifdef CONFIG_PKVM_INTEL_PVIOMMU
+	int order;
+#endif
 
 	if (iommu->qi_inited)
 		return 0;
@@ -289,15 +293,23 @@ static int create_qi_desc(struct pkvm_iommu *iommu)
 		return -EINVAL;
 
 	qi->free_cnt = PKVM_QI_DESC_ALIGNED_SIZE / sizeof(struct qi_desc);
+#ifdef CONFIG_PKVM_INTEL_PVIOMMU
+	qi->desc = pkvm_phys_to_virt(IQ_DESC_BASE_PHYS(iommu->viommu.iqa));
+	qi->desc_status = iommu_zalloc_pages(PKVM_QI_DESC_STATUS_ALIGNED_SIZE);
+	if (!qi->desc_status)
+		return -ENOMEM;
+	order = ecap_smts(iommu->iommu.ecap) ? 1 : 0;
+	__pkvm_host_donate_hyp(IQ_DESC_BASE_PHYS(iommu->viommu.iqa), (1 << order) * PAGE_SIZE);
+#else
 	qi->desc = iommu_zalloc_pages(PKVM_QI_DESC_ALIGNED_SIZE);
 	if (!qi->desc)
 		return -ENOMEM;
-
 	qi->desc_status = iommu_zalloc_pages(PKVM_QI_DESC_STATUS_ALIGNED_SIZE);
 	if (!qi->desc_status) {
 		iommu_put_page(qi->desc);
 		return -ENOMEM;
 	}
+#endif
 
 	enable_qi(iommu);
 	iommu->qi_inited = true;
@@ -422,6 +434,12 @@ static void submit_qi(struct pkvm_iommu *iommu, struct qi_desc *base, int count)
 		count -= submit_count;
 		base += submit_count;
 	} while (count > 0);
+}
+
+void pkvm_iommu_submit_qi(unsigned long phys, unsigned long desc_base, int count)
+{
+	struct pkvm_iommu *pkvm_iommu = find_iommu_by_reg_phys(phys);
+	submit_qi(pkvm_iommu, (struct qi_desc *)desc_base, count);
 }
 
 void flush_context_cache(struct pkvm_iommu *iommu, u16 did,
@@ -567,102 +585,6 @@ int activate_iommu(struct pkvm_iommu *iommu)
 
 free_shadow:
 	free_shadow_id(iommu, vaddr, vaddr_end);
-	return ret;
-}
-
-static void handle_qi_submit(struct pkvm_iommu *iommu, void *vdesc, int vhead, int count)
-{
-	struct pkvm_viommu *viommu = &iommu->viommu;
-	int vlen = IQ_DESC_LEN(viommu->iqa);
-	int vshift = IQ_DESC_SHIFT(viommu->iqa);
-	int len = IQ_DESC_LEN(iommu->piommu_iqa);
-	int shift = IQ_DESC_SHIFT(iommu->piommu_iqa);
-	struct q_inval *qi = &iommu->qi;
-	struct qi_desc *to, *from;
-	int required_cnt = count + 1, i;
-
-	pkvm_spin_lock(&iommu->qi_lock);
-	/*
-	 * Detect if the free descriptor count is enough or not
-	 */
-	while (qi->free_cnt < required_cnt) {
-		u64 head = readq(iommu->iommu.reg + DMAR_IQH_REG) >> shift;
-		int busy_cnt = (READ_ONCE(qi->free_head) + len - head) % len;
-		int free_cnt = len - busy_cnt;
-
-		if (free_cnt >= required_cnt) {
-			qi->free_cnt = free_cnt;
-			break;
-		}
-		pkvm_spin_unlock(&iommu->qi_lock);
-		cpu_relax();
-		pkvm_spin_lock(&iommu->qi_lock);
-	}
-
-	for (i = 0; i < count; i++) {
-		from = vdesc + (((vhead + i) % vlen) << vshift);
-		to = qi->desc + (((qi->free_head + i) % len) << shift);
-
-		to->qw0 = from->qw0;
-		to->qw1 = from->qw1;
-	}
-
-	/*
-	 * Reuse the desc_status from host so that host can poll
-	 * the desc_status itself instead of waiting in pkvm.
-	 */
-	qi->free_cnt -= count;
-	qi->free_head = (qi->free_head + count) % len;
-	writel(qi->free_head << shift, iommu->iommu.reg + DMAR_IQT_REG);
-
-	pkvm_spin_unlock(&iommu->qi_lock);
-}
-
-static int handle_qi_invalidation(struct pkvm_iommu *iommu, unsigned long val)
-{
-	struct pkvm_viommu *viommu = &iommu->viommu;
-	u64 viommu_iqa = viommu->iqa;
-	struct qi_desc *wait_desc;
-	int len = IQ_DESC_LEN(viommu_iqa);
-	int shift = IQ_DESC_SHIFT(viommu_iqa);
-	int head = viommu->vreg.iq_head >> shift;
-	int count, i, ret = 0;
-	int *desc_status;
-	void *desc;
-
-	viommu->vreg.iq_tail = val;
-	desc = pkvm_phys_to_virt(IQ_DESC_BASE_PHYS(viommu_iqa));
-	count = ((val >> shift) + len - head) % len;
-
-	for (i = 0; i < count; i++) {
-		viommu->vreg.iq_head = ((head + i) % len) << shift;
-		ret = handle_descriptor(iommu, desc + viommu->vreg.iq_head);
-		if (ret)
-			break;
-	}
-
-	/* update iq_head */
-	viommu->vreg.iq_head = val;
-
-	if (likely(!ret)) {
-		/*
-		 * Submit the descriptor to hardware. The desc_status
-		 * will be taken cared by hardware.
-		 */
-		handle_qi_submit(iommu, desc, head, count);
-	} else {
-		pkvm_err("pkvm: %s: failed with ret %d\n", __func__, ret);
-		/*
-		 * The descriptor seems invalid. Mark the desc_status as
-		 * QI_ABORT to make sure host driver won't be blocked.
-		 */
-		wait_desc = desc + (((head + count - 1) % len) << shift);
-		if (QI_DESC_TYPE(wait_desc->qw0) == QI_IWD_TYPE) {
-			desc_status = pkvm_phys_to_virt(wait_desc->qw1);
-			WRITE_ONCE(*desc_status, QI_ABORT);
-		}
-	}
-
 	return ret;
 }
 
@@ -1094,6 +1016,14 @@ void iommu_flush_iotlb(struct pkvm_iommu *iommu, struct iotlb_flush_data *data)
 	struct pkvm_ptdev *ptdev;
 	struct qi_desc *desc = data->desc;
 	int qi_desc_index = 0;
+
+	/*
+	 * QI Interface has not been intialized yet. Don't do anything.
+	 * TODO:
+	 * Put locking here.
+	 */
+	if (!iommu->piommu_iqa)
+		return;
 
 	pkvm_spin_lock(&iommu->lock);
